Assignment-1/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic-settings>=2.2.0
passlib[bcrypt]>=1.7.4
lxml>=5.2.0
dicttoxml>=1.7.4
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.2.0
flake8>=7.0.0
mypy>=1.10.0
